+++ src/components/utils.js
@@ -101,3 +101,25 @@ export const COLORMAP_OPTIONS = [
   'density',
   'inferno',
 ];
+
+export const DEFAULT_GL_OPTIONS = { webgl2: true };
+
+export function createDefaultUpdateStatus(componentName) {
+  return message => console.warn(`${componentName} updateStatus: ${message}`);
+}
+
+export function createDefaultUpdateCellsSelection(componentName) {
+  return cellsSelection => console.warn(`${componentName} updateCellsSelection: ${cellsSelection}`);
+}
+
+export function createDefaultUpdateCellsHover(componentName) {
+  return hoverInfo => console.warn(`${componentName} updateCellsHover: ${hoverInfo.cellId}`);
+}
+
+export function createDefaultUpdateViewInfo(componentName) {
+  return viewInfo => console.warn(`${componentName} updateViewInfo: ${viewInfo}`);
+}
+
+export function createDefaultClearPleaseWait(componentName) {
+  return layer => console.warn(`${componentName} "clearPleaseWait" not provided; layer: ${layer}`);
+}